requirements.txt
anthropic>=0.40.0
requests>=2.32.0
python-dotenv>=1.0.0
pdfplumber>=0.11.0
streamlit>=1.40.0
pytest>=8.0.0
